Complete event type select placement/styling
