+++ src/components/Dashboard/Dashboard.css
@@ -1,27 +1,30 @@
 .dashboard {
+  display: flex;
+  flex-direction: column;
   width: 100%;
-  padding: 16px 16px;
 }
 
 .dashboard-title {
-  margin-bottom: 24px;
-  padding-bottom: 8px;
-  border-bottom: solid 1px rgba(0, 0, 0, .2);
-  font-size: 1.50em;
-  text-align: center;
-  font-weight: 500;
+  height: 3.5rem;
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  margin-bottom: 20px;
+  padding-left: 20px;
+  border-bottom:solid 1px rgba(0, 0, 0, .1);
 }
 
 .dashboard-title > h1 {
-  font-size: inherit;
+  font-size: 1.5em;
+  font-weight: 600;
 }
 
 .event-type-select {
-  align-self: center;
+  align-self: flex-end;
   display: flex;
   align-items: center;
-  margin-bottom: 16px;
-  font-size: 1.25em;
+  margin-bottom: 24px;
+  padding-right: 20px;
   font-weight: 500;
 }
 
@@ -29,16 +32,18 @@
   background-color: unset;
   padding: unset;
   border: unset;
-  font-size: inherit;
+  font-size: 1.15em;
   font-weight: inherit;
+  cursor: pointer;
 }
 
 .event-type-btn:nth-child(1) {
-  margin-right: 8px;
+  padding-right: 8px;
+  border-right: solid 1px rgba(0, 0, 0, .2);
 }
 
 .event-type-btn:nth-child(2) {
-  margin-left: 8px;
+  padding-left: 8px;
 }
 
 .unselected {
@@ -54,10 +59,10 @@
 @media only screen and (min-width: 768px) {
   .dashboard {
     height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding: 20px;
+  }
+
+  .dashboard-title {
+    margin-bottom: 32px;
   }
 
   .event-type-select {
@@ -71,5 +76,6 @@
     grid-template-rows: unset;
     column-gap: 20px;
     margin: 0 auto;
+    padding: 0 20px;
   }
 }
