+++ src/components/TeeTimeContainer/InviteTypeSelect/InviteTypeSelect.css
@@ -1,0 +1,33 @@
+.invite-type-select {
+  align-self: flex-start;
+  margin: 0 12px 12px 0;
+}
+
+.type-select-btn {
+  background-color: unset;
+  padding: unset;
+  border: unset;
+  font-size: 1.15em;
+  font-weight: 500;
+  cursor: pointer;
+}
+
+.type-select-btn:nth-child(1) {
+  padding-right: 6px;
+  border-right: solid 1px rgba(0, 0, 0, .2);
+}
+
+.type-select-btn:nth-child(2) {
+  padding-left: 6px;
+}
+
+.unselected {
+  color: rgba(0, 0, 0, .25);
+}
+
+@media only screen and (min-width: 768px) {
+  .invite-type-select {
+    align-self: unset;
+    margin: unset;
+  }
+}
