+++ src/components/TeeTimeContainer/TeeTimeContainer.css
@@ -2,6 +2,7 @@
   width: 100%;
   display: flex;
   flex-direction: column;
+  padding: 0 20px;
 }
 
 .container-title {
@@ -10,7 +11,7 @@
 
 .container-title > h2 {
   font-size: 1.25em;
-    font-weight: 500;
+  font-weight: 600;
 }
 
 .type-select-con {
@@ -29,6 +30,7 @@
     grid-template-rows: 4.5rem auto;
     grid-auto-rows: 3.75rem;
     margin-bottom: unset;
+    padding: unset;
     border-radius: 16px;
     box-shadow: 0 1px 4px rgba(0, 0, 0, .2);
   }
